# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: IPRS MVP API
  version: 0.1.0
servers:
  - url: http://localhost:8000
...
paths:
...
  /api/v1/marketdata/snapshots:
    post:
      tags: [MarketData]
      summary: Create market data snapshot
      operationId: createMarketDataSnapshot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarketDataSnapshotV1"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  snapshot_id: { type: string }
                required: [snapshot_id]
  /api/v1/marketdata/snapshots/{snapshotId}:
    get:
      tags: [MarketData]
      summary: Get market data snapshot by ID
      operationId: getMarketDataSnapshot
      parameters:
        - in: path
          name: snapshotId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketDataSnapshotV1"
        "404":
          description: Not found
...
  /api/v1/positions:
    get:
      tags: [Positions]
      summary: Query positions for a snapshot scope
      operationId: getPositions
      parameters:
        - in: query
          name: asOf
          required: true
          schema: { type: string, format: date-time }
        - in: query
          name: portfolioNodeId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: PositionSnapshotV1
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PositionSnapshotV1"
...
  /api/v1/runs:
    post:
      tags: [Runs]
      summary: Create a run and enqueue tasks
      operationId: createRun
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunRequestedV1"
      responses:
        "201":
          description: Run created
          content:
            application/json:
              schema:
                type: object
                properties:
                  run_id: { type: string }
                  status: { type: string }
                required: [run_id, status]
    get:
      tags: [Runs]
      summary: List runs
      operationId: listRuns
      parameters:
        - in: query
          name: status
          required: false
          schema: { type: string }
        - in: query
          name: type
          required: false
          schema: { type: string }
        - in: query
          name: from
          required: false
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: Runs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RunRecord"
...
components:
  schemas:
    MarketDataSnapshotV1:
      type: object
      required: [snapshot_id, as_of_time, vendor, universe_id, fx_spots, curves, quality]
      properties:
        snapshot_id: { type: string }
        as_of_time: { type: string, format: date-time }
        vendor: { type: string }
        universe_id: { type: string }
        fx_spots:
          type: array
          items:
            type: object
            required: [pair, spot, ts]
            properties:
              pair: { type: string }
              spot: { type: number }
              ts: { type: string, format: date-time }
        curves:
          type: array
          items:
            type: object
            required: [curve_id, curve_type, nodes]
            properties:
              curve_id:
                type: string
                enum: [USD-OIS, EUR-OIS, LOAN-SPREAD, FI-SPREAD]
              curve_type:
                type: string
                enum: [DISCOUNT, SPREAD]
              nodes:
                type: array
                items:
                  type: object
                  required: [tenor, zero_rate]
                  properties:
                    tenor: { type: string }
                    zero_rate: { type: number }
        quality:
          type: object
          required: [dq_status, issues]
          properties:
            dq_status: { type: string, enum: [PASS, WARN, FAIL] }
            issues:
              type: array
              items: { type: string }
    PositionSnapshotV1:
      type: object
      required: [position_snapshot_id, as_of_time, portfolio_node_id, positions]
      properties:
        position_snapshot_id: { type: string }
        as_of_time: { type: string, format: date-time }
        portfolio_node_id: { type: string }
        positions:
          type: array
          items:
            type: object
            required: [position_id, instrument_id, product_type, quantity]
            properties:
              position_id: { type: string }
              instrument_id: { type: string }
              product_type: { type: string, enum: [FX_FWD, AMORT_LOAN, FIXED_BOND] }
              quantity: { type: number }
              attributes: { type: object }
...
    RunRequestedV1:
      type: object
      required: [run_id, run_type, as_of_time, market_snapshot_id, measures, scenarios, portfolio_scope]
      properties:
        run_id: { type: string }
        run_type: { type: string, enum: [EOD_OFFICIAL, INTRADAY, SANDBOX] }
        as_of_time: { type: string, format: date-time }
        market_snapshot_id: { type: string }
        model_set_id: { type: string }
        portfolio_scope:
          type: object
          required: [node_ids]
          properties:
            node_ids:
              type: array
              items: { type: string }
        measures:
          type: array
          items: { type: string }
        scenarios:
          type: array
          items: { type: object }
        execution: { type: object }
    RunRecord:
      type: object
      required: [run_id, run_type, status, requested_at, as_of_time, market_snapshot_id, measures]
      properties:
        run_id: { type: string }
        run_type: { type: string }
        status: { type: string }
        requested_by: { type: string }
        requested_at: { type: string, format: date-time }
        as_of_time: { type: string, format: date-time }
        market_snapshot_id: { type: string }
        model_set_id: { type: string }
        measures:
          type: array
          items: { type: string }
        scenarios: { type: object }
        portfolio_scope: { type: object }
        started_at: { type: string, format: date-time }
        completed_at: { type: string, format: date-time }
        summary_json: { type: object }
        error_json: { type: object }
...